pyproject.toml
[tool.poetry]
name = "novatus"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = true
packages = [{include = "novatus_cli",from = "src"}]


[tool.poetry.scripts]
novatus = "novatus_cli.main:app"

[tool.poetry.dependencies]
python = "^3.11"
pre-commit = "^4.1.0"
psycopg2-binary = "^2.9.10"
typer = "^0.15.1"
docker = "^7.1.0"
sqlalchemy = "^2.0.38"
pandas = "2.1.4"
pyarrow = "^19.0.1"
pyyaml = "^6.0.2"
faker = "^37.3.0"
fastapi = "^0.115.12"
uvicorn = "^0.34.2"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.2"


[tool.poetry.group.dev.dependencies]
ruff = "^0.5.7"
mypy = "^1.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "N",  # PEP8 naming conventions
    "D" # Docstrings
]
# Rules can be found https://docs.astral.sh/ruff/rules/
ignore = [ # As per https://docs.astral.sh/ruff/formatter/#format-suppression
    "D103",
    "ISC002",
    "ISC001",
    "COM819",
    "COM812",
    "Q003",
    "Q002",
    "Q001",
    "Q000",
    "D300",
    "D206",
    "E117",
    "E114",
    "E111",
    "W191",
    "D105",
    "D203", # incompatible with D211
    "D213", # incompatible with D212,
    "D401",
    "D205",
    "D104",
    "D400",
    "D415",
    "D102",
    "D107",
    "D102"
]

[tool.pytest.ini_options]
pythonpath = [
  "src"
]
testpaths = ["tests","src"]